requirements.txt
langchain==0.3.21
langchain_community==0.3.20
langchain_chroma==0.2.2
langchain-ollama==0.3.0
chromadb==0.6.3
pydantic<3.0.0,>=2.7.4
SQLAlchemy<3,>=1.4
requests<3,>=2
PyYAML>=5.3
aiohttp<4.0.0,>=3.8.3
tenacity!=8.4.0,<10,>=8.1.0
dataclasses-json<0.7,>=0.5.7
pydantic-settings<3.0.0,>=2.4.0
httpx-sse<1.0.0,>=0.4.0
numpy<3,>=1.26.2
ollama<1,>=0.4.4
build>=1.0.3
chroma-hnswlib==0.7.6
fastapi>=0.95.2
uvicorn>=0.18.3
posthog>=2.4.0
onnxruntime>=1.14.1
opentelemetry-api>=1.2.0
opentelemetry-exporter-otlp-proto-grpc>=1.2.0
opentelemetry-instrumentation-fastapi>=0.41b0
pypika>=0.48.9
importlib-resources
bcrypt>=4.0.1
typer>=0.9.0
kubernetes>=28.1.0
mmh3>=4.0.1
httpx>=0.27.0
rich>=10.11.0
sentence-transformers
pytubefix
pydub
python-dotenv
openai-whisper
gradio